Replace unsigned long long int with uint64_t for priority values
uint64_t from <stdint.h> is exactly 64 bits on all platforms, making the
intent precise. Replaces <limits.h>/ULONG_MAX with <stdint.h>/UINT64_MAX
throughout the header, implementation, and tests.

+++ include/priority_queue.h
@@ -5,6 +5,7 @@
 
 #include <stdbool.h>
 #include <stddef.h>
+#include <stdint.h>
 
 /**
  * How to get the priority out of the generic element
@@ -14,11 +15,11 @@
  * @param element
  * @returns priority (a u64)
  **/
-typedef unsigned long long int (*priority_queue_get_priority_t)(void *element);
+typedef uint64_t (*priority_queue_get_priority_t)(void *element);
 
 struct priority_queue {
 	// We assume that priority is expressed in terms of a 64 bit unsigned integral
-	unsigned long long int        highest_priority;
+	uint64_t        highest_priority;
 	void                         *items[MAX];
 	size_t                        first_free;
 	priority_queue_get_priority_t get_priority;
